requirements.txt
fastapi
pandas
python-dotenv
openai
uvicorn
